fixing some errors and details
Placing albums that didn't have categories in later identified category folders.
Fixing some errors that occurred when inserting a category link and album link belonging to that category together to execute.
Solved error in find_key, where when looking for values ​​in the past dictionary, the dictionary elements were removed. Now a deepcopy is done so that it doesn't interfere with the root dictionary.